Support spaces in filenames
Some of my directories contained spaces and the script failed on that, as the for loop considered the parts of the filename separately.
This change improves quoting and used the mapfile command (introduced in Bash version 4). I briefly tried setting the Internal Field Separator `$IFS` but couldn't get this to quickly, so I decided to disregard backwards compatibility and use the mapfile command.